+++ challenge3/webapp/hiroaki_terayama/app/controllers/admins/energy_records_controller.rb
@@ -1,0 +1,16 @@
+module Admins
+  class EnergyRecordsController < ApplicationController
+
+    def new
+    end
+
+    def create
+      errors = EnergyRecord.import_csv(params[:csv_file].path)
+      if !errors
+        redirect_to new_energy_record_path, flash: { success: 'インポート成功！' }
+      else
+        render :new
+      end
+    end
+  end
+end
