+++ challenge3/webapp/hiroaki_terayama/app/controllers/admins_controller.rb
@@ -1,0 +1,5 @@
+class AdminsController < ApplicationController
+
+  def index
+  end
+end
