+++ challenge3/webapp/hiroaki_terayama/app/views/admins/energy_records/new.html.erb
@@ -1,5 +1,5 @@
-<div class="subtitle">house_dataインポート</div>
-<%= form_with url: houses_path, method: :post, multipart: true do |f| %>
+<div class="subtitle">dataset_50インポート</div>
+<%= form_with url: admins_energy_records_path, method: :post, multipart: true do |f| %>
   <%= flash[:success] %>
   <div class="field">
     <%= f.label "CSVファイルを選択" %><br>
@@ -9,5 +9,5 @@
   <div class="actions">
     <%= f.submit "CSVファイルを読み込む", class: "button is-primary is-text-weight-bold"%>
   </div>
-  <%= link_to '→dataset_50をインポートする', new_energy_record_path %>
+  <%= link_to '→house_dataをインポートする', new_admins_house_path %>
 <% end %>
