+++ challenge3/webapp/hiroaki_terayama/app/views/admins/index.html.erb
@@ -1,0 +1,2 @@
+<%= link_to '→house_dataをインポートする', new_admins_house_path %><br>
+<%= link_to '→dataset_50をインポートする', new_admins_energy_record_path %>
